+++ include/proto/block.h
@@ -18,12 +18,10 @@ namespace proto {
     public:
         class Header {
         public:
-            // current block number
             BlockNumber number{};
-            // previous block header hash
+            // previous hash of ALL the user request
             HashString previousHash{};
-            // current block body (user request) hash
-            // exclude the execution result
+            // current user request hash
             HashString dataHash{};
 
         public:
@@ -65,20 +63,6 @@ namespace proto {
             constexpr static auto serialize(auto &archive, Body &b) {
                 return archive(b.userRequests);
             }
-
-            [[nodiscard]] bool serializeForProofGen(std::vector<int>& posList, std::string& ret, int startPos = 0) const {
-                posList.resize(userRequests.size());
-                zpp::bits::out out(ret);
-                out.reset(startPos);
-                for (int i=0; i<(int)userRequests.size(); i++) {
-                    // serialize std::unique_ptr<Envelop>
-                    if(failure(out(userRequests[i]))) {
-                        return false;
-                    }
-                    posList[i] = (int)out.position();
-                }
-                return true;
-            }
         };
 
         class ExecuteResult {
@@ -93,20 +77,6 @@ namespace proto {
             constexpr static auto serialize(auto &archive, ExecuteResult &e) {
                 return archive(e.txReadWriteSet, e.transactionFilter);
             }
-
-            [[nodiscard]] bool serializeForProofGen(std::vector<int>& posList, std::string& ret, int startPos = 0) const {
-                posList.resize(txReadWriteSet.size());
-                zpp::bits::out out(ret);
-                out.reset(startPos);
-                for (int i=0; i<(int)txReadWriteSet.size(); i++) {
-                    // serialize std::unique_ptr<Envelop>
-                    if(failure(out(txReadWriteSet[i], transactionFilter[i]))) {
-                        return false;
-                    }
-                    posList[i] = (int)out.position();
-                }
-                return true;
-            }
         };
 
         // std::string: the metadata to be signed (may leave empty)
@@ -115,9 +85,9 @@ namespace proto {
 
         class Metadata {
         public:
-            // when a peer received a block after BFT consensus, it adds the results to signatures.
+            // when a peer validated a block(before execution), it adds its signature of the HEADER+BODY to signatures.
             std::vector<SignaturePair> consensusSignatures;
-            // when a peer executed a block, it adds its signature of the (HEADER+ExecuteResult) to signatures.
+            // when a peer validated a block(after execution), it adds its signature of the (HEADER+BODY+ExecuteResult) to signatures.
             std::vector<SignaturePair> validateSignatures;
         public:
             friend zpp::bits::access;
